+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/apache/ApacheRestCallFactory.java
@@ -11,6 +11,7 @@ import at.porscheinformatik.happyrest.RestLoggerAdapter;
 import at.porscheinformatik.happyrest.RestParser;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedFormatter;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedParser;
+import at.porscheinformatik.happyrest.util.BinaryParser;
 import at.porscheinformatik.happyrest.util.TextPlainFormatter;
 
 /**
@@ -24,7 +25,7 @@ public class ApacheRestCallFactory implements RestCallFactory
     public static ApacheRestCallFactory create(RestLoggerAdapter loggerAdapter, ObjectMapper mapper)
     {
         RestFormatter formatter = RestFormatter.of(new JacksonBasedFormatter(mapper), new TextPlainFormatter());
-        RestParser parser = RestParser.of(new JacksonBasedParser(mapper));
+        RestParser parser = RestParser.of(new JacksonBasedParser(mapper), BinaryParser.INSTANCE);
 
         return new ApacheRestCallFactory(loggerAdapter, formatter, parser);
     }